requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.4.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.1
cucim-cu12==24.6.0
cuda-python==12.5.0
cudf-cu12==24.6.1
cugraph-cu12==24.6.1
cuml-cu12==24.6.1
cuproj-cu12==24.6.0
cupy-cuda12x==13.2.0
cuspatial-cu12==24.6.0
cuvs-cu12==24.6.0
cuxfilter-cu12==24.6.0
cycler==0.12.1
dask==2024.5.1
dask-cuda==24.6.0
dask-cudf-cu12==24.6.1
dask-expr==1.1.1
datashader==0.16.3
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.5.1
distributed-ucxx-cu12==0.38.0
executing==2.0.1
fastjsonschema==2.20.0
fastrlock==0.8.2
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
geopandas==1.0.1
holoviews==1.19.1
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_proxy==4.3.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
libucx-cu12==1.15.0.post1
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvtx==0.2.10
openTSNE==1.0.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
panel==1.4.4
param==2.1.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.10.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.4.0
pyarrow==16.1.0
pycparser==2.22
pyct==0.5.0
Pygments==2.18.0
pylibcugraph-cu12==24.6.1
pylibraft-cu12==24.6.0
pynvjitlink-cu12==0.2.4
pynvml==11.4.1
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
pyzmq==26.0.3
raft-dask-cu12==24.6.0
rapids-dask-dependency==24.6.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rmm-cu12==24.6.0
rpds-py==0.18.1
SciencePlots==2.1.1
scikit-image==0.23.2
scikit-learn==1.5.1
scikit-optimize==0.10.2
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.0.4
simpervisor==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
tblib==3.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
treelite==4.1.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ucx-py-cu12==0.38.0
ucxx-cu12==0.38.0
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
xarray==2024.6.0
xyzservices==2024.6.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2
